Implement adaptive indexer
Index markdown docs up to H3 if token size exceeeds the given max token size
