+++ doc_indexer/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "indexing"
+name = "doc_indexer"
 version = "0.1.0"
 description = "Indexes Kyma documents for RAG"
 authors = []
@@ -9,7 +9,7 @@ package-mode = false
 python = "~3.12"
 hdbcli = "^2.22.29"
 python-decouple = "^3.8"
-generative-ai-hub-sdk = {extras = ["all"], version = "^3.2.6"}
+generative-ai-hub-sdk = {extras = ["all"], version = "^4.0.0"}
 python-dotenv = "^1.0.1"
 unstructured = "^0.16.0"
 markdown = "^3.7"
