+++ doc_indexer/src/indexing/contants.py
@@ -1,2 +1,3 @@
 HEADER1 = ("#", "Header1")
 HEADER2 = ("##", "Header2")
+HEADER3 = ("###", "Header3")
